+++ packages/contracts/lib.esm/index.js
@@ -712,7 +712,7 @@ export class BaseContract {
             delete this._runningEvents[runningEvent.tag];
             // If we have a poller for this, remove it
             const emit = this._wrappedEmits[runningEvent.tag];
-            if (emit) {
+            if (emit && runningEvent.filter) {
                 this.provider.off(runningEvent.filter, emit);
                 delete this._wrappedEmits[runningEvent.tag];
             }