+++ packages/ethers/lib.esm/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "ethers/5.1.1";
+export declare const version = "ethers/5.1.2";
 //# sourceMappingURL=_version.d.ts.map